LPS-153305 Move _getDefaultLiferayHome() from PropsUtil to SystemProperties, move the properties 'default.liferay.home
